builder/libvirt: Get VM up and running
We now start the VM, and type the boot command.  The next thing to get
sorted is the network - we need to know the IP of the VM, and the IP
of the host that the VM can see.

Probably the best thing is to create a new libvirt network to connect
the machine to?